requirements.txt
nextcord==2.4.2
openai==0.27.2
pymongo==4.3.3
python-dotenv==1.0.0